+++ orca/__init__.py
@@ -1,3 +1,5 @@
+__version__ = "1.9.dev0"
+
 from .orca import *
 
-version = __version__ = '1.8'
+version = __version__