+++ orca/utils/testing.py
@@ -6,7 +6,6 @@
 Utilities used in testing of Orca.
 
 """
-import numpy as np
 import numpy.testing as npt
 import pandas as pd
 
@@ -42,7 +41,7 @@ def assert_frames_equal(actual, expected, use_close=False):
 
         act_row = actual.loc[i]
 
-        for j, exp_item in exp_row.iteritems():
+        for j, exp_item in exp_row.items():
             assert j in act_row.index, \
                 'Expected column {!r} not found.'.format(j)